requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.8
openai==1.3.7
google-generativeai==0.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
